fix(modules): improve Windows runtime portability
Replace GNU-only cases and POSIX assumptions with portable PX4 helpers across drivers, examples, logger, dataman, version reporting, system commands, and tests.

Keep ioctl argument types consistent and make startup/runtime diagnostics less noisy on SITL-only platforms where optional devices are absent.

+++ src/drivers/px4io/px4io.cpp
@@ -130,7 +130,7 @@ public:
 	 * @param[in] cmd the IOCTL command
 	 * @param[in] the IOCTL command parameter (optional)
 	 */
-	int		ioctl(file *filp, int cmd, unsigned long arg) override;
+	int		ioctl(file *filp, int cmd, uintptr_t arg) override;
 
 	/**
 	 * Print IO status.
@@ -1309,7 +1309,7 @@ int PX4IO::print_status()
 	return 0;
 }
 
-int PX4IO::ioctl(file *filep, int cmd, unsigned long arg)
+int PX4IO::ioctl(file *filep, int cmd, uintptr_t arg)
 {
 	SmartLock lock_guard(_lock);
 	int ret = OK;
