+++ src/lib/cdev/CDev.hpp
@@ -146,7 +146,7 @@ public:
 	 * @param arg		The ioctl argument value.
 	 * @return		OK on success, or -errno otherwise.
 	 */
-	virtual int	ioctl(file_t *filep, int cmd, unsigned long arg) { return -ENOTTY; };
+	virtual int	ioctl(file_t *filep, int cmd, uintptr_t arg) { return -ENOTTY; };
 
 	/**
 	 * Perform a poll setup/teardown operation.
